refactor: add missing ones in previous commit
- missed 3 commits in previous commit... adding!
