--- message.js
@@ -1,9 +1,0 @@
-const message = {
-    INPUT_STATION_NAME_LESS_THAN_TWO : "역 이름은 2글자 이상이여야합니다.",
-    INPUT_STATION_NAME_OVERLAPPED : "중복된 역이 있습니다.",
-    DELETE_STATION_BTN_PRESSED : "선택하신 역을 삭제하시겠습니까?",
-    LINE_NAME_FORMAT_INAPPROPRIATE : "◯호선 <-- 형식으로 입력해주세요.",
-    REGISTER_LINE_OVERLAPPED : "상행종점과 하행종점이 같습니다. 다르게 설정해야 합니다."
-};
-
-export {message};
